request trace: trim agent module leftovers
