+++ lib/llm/src/protocols/openai/nvext.rs
@@ -5,11 +5,10 @@ use axum::http::HeaderMap;
 use derive_builder::Builder;
 use dynamo_kv_router::protocols::RoutingConstraints;
 use dynamo_protocols::types::StopReason;
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Deserializer, Serialize};
 use utoipa::ToSchema;
 use validator::{Validate, ValidationError};
 
-pub use crate::agents::context::AgentContext;
 use crate::protocols::TokenIdType;
 pub use crate::protocols::common::llm_backend::PromptLogprobs;
 pub use crate::protocols::common::timing::TimingInfo;
@@ -409,6 +408,53 @@ where
         ));
     }
     Ok(url.to_string())
+}
+
+fn deserialize_non_empty_agent_context_string<'de, D>(deserializer: D) -> Result<String, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let value = String::deserialize(deserializer)?;
+    if value.trim().is_empty() {
+        return Err(serde::de::Error::custom(
+            "agent_context required identifiers must be non-empty",
+        ));
+    }
+    Ok(value)
+}
+
+/// Identity metadata for agentic workloads.
+#[derive(ToSchema, Serialize, Deserialize, Builder, Debug, Clone, PartialEq, Eq)]
+pub struct AgentContext {
+    /// Reusable session/profile class.
+    #[serde(deserialize_with = "deserialize_non_empty_agent_context_string")]
+    pub session_type_id: String,
+
+    /// Top-level agent run/session identifier.
+    #[serde(deserialize_with = "deserialize_non_empty_agent_context_string")]
+    pub session_id: String,
+
+    /// Schedulable reasoning/tool trajectory identifier.
+    #[serde(deserialize_with = "deserialize_non_empty_agent_context_string")]
+    pub trajectory_id: String,
+
+    /// Optional parent trajectory for subagents.
+    #[builder(default, setter(strip_option))]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub parent_trajectory_id: Option<String>,
+
+    /// Optional terminal marker: when true, this request signals that the
+    /// trajectory is complete. Lifecycle-aware backends use it to release any
+    /// per-trajectory state they hold right away; other backends ignore it.
+    #[builder(default, setter(strip_option))]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub trajectory_final: Option<bool>,
+}
+
+impl AgentContext {
+    pub fn builder() -> AgentContextBuilder {
+        AgentContextBuilder::default()
+    }
 }
 
 /// NVIDIA LLM extensions to the OpenAI API
@@ -816,6 +862,21 @@ mod tests {
         }"#;
 
         assert!(serde_json::from_str::<NvExt>(json).is_err());
+    }
+
+    #[test]
+    fn test_agent_context_deserialize_rejects_empty_required_ids() {
+        let err = serde_json::from_value::<AgentContext>(serde_json::json!({
+            "session_type_id": "deep_research",
+            "session_id": "",
+            "trajectory_id": "trajectory-1"
+        }))
+        .expect_err("empty session_id should fail deserialization");
+
+        assert!(
+            err.to_string()
+                .contains("agent_context required identifiers must be non-empty")
+        );
     }
 
     #[test]
