+++ lib/llm/src/request_trace/agent_context.rs
@@ -13,13 +13,13 @@ use dynamo_runtime::protocols::annotated::Annotated;
 use dynamo_runtime::transports::event_plane::EventSubscriber;
 use parking_lot::Mutex;
 
-use crate::agents::context::AgentContext;
 use crate::local_model::LocalModel;
 use crate::protocols::common::FinishReason as BackendFinishReason;
 use crate::protocols::common::preprocessor::PreprocessedRequest;
 use crate::protocols::common::timing::RequestTracker;
 use crate::protocols::openai::{
-    chat_completions::NvCreateChatCompletionStreamResponse, completions::NvCreateCompletionResponse,
+    chat_completions::NvCreateChatCompletionStreamResponse,
+    completions::NvCreateCompletionResponse, nvext::AgentContext,
 };
 use crate::request_trace::{
     DEFAULT_TOOL_EVENTS_TOPIC, FinishReasonMetadata, RequestReplayMetrics, RequestTraceMetrics,
@@ -523,14 +523,13 @@ fn tool_events_namespace(local_model: &LocalModel) -> String {
 mod tests {
     use std::{thread, time::Duration};
 
-    use crate::agents::context::AgentContext;
     use crate::protocols::common::{
         self,
         timing::{RequestTracker, WORKER_TYPE_DECODE},
     };
     use crate::protocols::openai::{
         chat_completions::NvCreateChatCompletionStreamResponse,
-        completions::NvCreateCompletionResponse,
+        completions::NvCreateCompletionResponse, nvext::AgentContext,
     };
     use dynamo_protocols::types::{
         ChatChoiceStream, ChatCompletionMessageToolCallChunk, ChatCompletionStreamResponseDelta,
